backend: Address review feedback for headlampconfig tests
- Move SetBaseLoggerForTest to export_test.go to avoid production overhead
- Add missing field assertions (NodeShellNamespace, ClusterInventoryProviderFile, ClusterInventoryLabelSelector)
- Remove unused getBaseLogger and use zlog.Logger directly in production code
- Remove unnecessary gosec nolint that wasn't triggering

All tests passing and linter clean.
