+++ backend/pkg/logger/export_test.go
@@ -1,0 +1,37 @@
+/*
+Copyright 2025 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package logger
+
+import (
+	"sync"
+
+	"github.com/rs/zerolog"
+	zlog "github.com/rs/zerolog/log"
+)
+
+var baseLoggerMutex sync.RWMutex
+
+// SetBaseLoggerForTest sets the base logger for testing and returns the previous one.
+func SetBaseLoggerForTest(l zerolog.Logger) zerolog.Logger {
+	baseLoggerMutex.Lock()
+	defer baseLoggerMutex.Unlock()
+
+	old := zlog.Logger
+	zlog.Logger = l
+
+	return old
+}
